Emulate 0-sized CDROM DMAs correctly
Ape Escape does not current work. This is due to some weird behaviour in the game itself; namely that it performs a 0-sized CDROM DMA (and expects it to transfer something!). To elaborate:

The game correctly locates a file on disc (DEMO/TITLE.ALL;1), commands the CDROM to read it, and then performs a 0-sized CDROM DMA (ie. it writes 0x00010000 to D3_BCR) to read the file's header. From this it expects to see at least 0x24B of data at the desired address, which it then uses to find the size of the file. Since this is (incorrectly) all 0 bytes, it reads a size of 0,
and the loader routine fails due to underflowing the size, causing it to try to load 4GB of data into the PSX memory.

On real PSX hardware, performing such a 0-sized CDROM (or OTC) DMA causes it to read 0x10000 words (0x40000B), not 0! The data after the sector is padded with a byte from the first sector. This has been verified on real hardware.

This commit makes psxjin correctly emulate this behaviour, making Ape Escape work correctly.

This might fix other games too, and hopefully won't break any others...

Thanks to Martin Korth (nocash.emubase.de) for finding out the behaviour of these DMAs on real hardware!